Added Handler

+++ payment_service/internal/models/payment_info.go
@@ -4,6 +4,6 @@ import "github.com/google/uuid"
 
 type PaymentInfo struct {
 	LessonID       uuid.UUID
-	PriceRUB       uint64
+	PriceRUB       int32
 	PaymentDetails string
 }
